Separate notification sound in notification_sound.go file
