+++ main.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"embed"
 	"flag"
 	"fmt"
 	"log"
@@ -9,17 +8,9 @@ import (
 	"time"
 
 	"gioui.org/x/notify"
-	"github.com/gopxl/beep"
-	"github.com/gopxl/beep/speaker"
-	"github.com/gopxl/beep/vorbis"
 )
 
-var (
-	//go:embed notification.ogg
-	NotificationSound embed.FS
-	Buffer            *beep.Buffer
-	Version           = "development"
-)
+var version = "development"
 
 type flags struct {
 	durationInSec  uint
@@ -50,15 +41,6 @@ func parseFlags() flags {
 		frequencyInMin: *frequencyInMin,
 		version:        *version,
 	}
-}
-
-func playNotificationSound() {
-	done := make(chan bool)
-	speaker.Play(
-		beep.Seq(Buffer.Streamer(0, Buffer.Len())),
-		beep.Callback(func() { done <- true }),
-	)
-	<-done
 }
 
 func sendNotification(notifier notify.Notifier, title string, text string) notify.Notification {
@@ -99,27 +81,10 @@ func twentyTwentyTwenty(notifier notify.Notifier, duration time.Duration, freque
 	}
 }
 
-func init() {
-	f, err := NotificationSound.Open("notification.ogg")
-	if err != nil {
-		log.Fatalf("Failed to load notification sound: %v\n", err)
-	}
-
-	streamer, format, err := vorbis.Decode(f)
-	if err != nil {
-		log.Fatalf("Failed to decode the notification sound: %v\n", err)
-	}
-	Buffer = beep.NewBuffer(format)
-	Buffer.Append(streamer)
-
-	// 1s/4 = 250ms of lag, good enough for this use case
-	speaker.Init(format.SampleRate, format.SampleRate.N(time.Second/4))
-}
-
 func main() {
 	flags := parseFlags()
 	if flags.version {
-		fmt.Println(Version)
+		fmt.Println(version)
 		os.Exit(0)
 	}
